Split up cli.js tests
Tests from this file are now consistently timing out in TravisCI. Split
across multiple files in the hope that this reduces the failure rate.
